API-4610: replace whitelist with allowlist, and mandate IP allowlist …

+++ app/service/IpAllowlistService.scala
@@ -23,7 +23,7 @@ import domain.models.flows.IpAllowlistFlow
 import javax.inject.{Inject, Singleton}
 import repositories.FlowRepository
 import repositories.ReactiveMongoFormatters.formatIpAllowlistFlow
-import uk.gov.hmrc.http.HeaderCarrier
+import uk.gov.hmrc.http.{ForbiddenException, HeaderCarrier}
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -33,7 +33,7 @@ class IpAllowlistService @Inject()(flowRepository: FlowRepository, connectorWrap
   def getIpAllowlistFlow(app: Application, sessionId: String): Future[IpAllowlistFlow] = {
     for {
       optionalFlow <- flowRepository.fetchBySessionIdAndFlowType[IpAllowlistFlow](sessionId, IP_ALLOW_LIST)
-      flow = optionalFlow.getOrElse(IpAllowlistFlow(sessionId, app.ipWhitelist))
+      flow = optionalFlow.getOrElse(IpAllowlistFlow(sessionId, app.ipAllowlist.allowlist))
       savedFlow <- flowRepository.saveFlow(flow)
     } yield savedFlow
   }
@@ -45,7 +45,7 @@ class IpAllowlistService @Inject()(flowRepository: FlowRepository, connectorWrap
   def addCidrBlock(cidrBlock: String, app: Application, sessionId: String): Future[IpAllowlistFlow] = {
     for {
       optionalFlow <- flowRepository.fetchBySessionIdAndFlowType[IpAllowlistFlow](sessionId, IP_ALLOW_LIST)
-      updatedFlow = optionalFlow.fold(IpAllowlistFlow(sessionId, app.ipWhitelist + cidrBlock))(flow => flow.copy(allowlist = flow.allowlist + cidrBlock))
+      updatedFlow = optionalFlow.fold(IpAllowlistFlow(sessionId, app.ipAllowlist.allowlist + cidrBlock))(flow => flow.copy(allowlist = flow.allowlist + cidrBlock))
       savedFlow <- flowRepository.saveFlow(updatedFlow)
     } yield savedFlow
   }
@@ -62,15 +62,20 @@ class IpAllowlistService @Inject()(flowRepository: FlowRepository, connectorWrap
     for {
       optionalFlow <- flowRepository.fetchBySessionIdAndFlowType[IpAllowlistFlow](sessionId, IP_ALLOW_LIST)
       flow = optionalFlow.getOrElse(throw new IllegalStateException(s"No IP allowlist flow exists for session ID $sessionId"))
-      result <- connectorWrapper.forEnvironment(app.deployedTo).thirdPartyApplicationConnector.updateIpAllowlist(app.id, flow.allowlist)
+      _ = if (flow.allowlist.isEmpty) throw new ForbiddenException(s"IP allowlist for session ID $sessionId cannot be activated because it is empty")
+      result <- connectorWrapper.forEnvironment(app.deployedTo).thirdPartyApplicationConnector.updateIpAllowlist(app.id, app.ipAllowlist.required, flow.allowlist)
       _ <- flowRepository.deleteBySessionIdAndFlowType(sessionId, IP_ALLOW_LIST)
     } yield result
   }
 
   def deactivateIpAllowlist(app: Application, sessionId: String)(implicit hc: HeaderCarrier): Future[ApplicationUpdateSuccessful] = {
-    for {
-      result <- connectorWrapper.forEnvironment(app.deployedTo).thirdPartyApplicationConnector.updateIpAllowlist(app.id, Set.empty)
-      _ <- flowRepository.deleteBySessionIdAndFlowType(sessionId, IP_ALLOW_LIST)
-    } yield result
+    if (app.ipAllowlist.required) {
+      Future.failed(new ForbiddenException(s"IP allowlist for session ID $sessionId cannot be deactivated because it is required"))
+    } else {
+      for {
+        result <- connectorWrapper.forEnvironment(app.deployedTo).thirdPartyApplicationConnector.updateIpAllowlist(app.id, app.ipAllowlist.required, Set.empty)
+        _ <- flowRepository.deleteBySessionIdAndFlowType(sessionId, IP_ALLOW_LIST)
+      } yield result
+    }
   }
 }
